double to long

+++ flink-experiments/src/main/java/org/apache/flink/experiments/DolToEur.java
@@ -3,7 +3,7 @@ package org.apache.flink.experiments;
 import org.apache.flink.table.functions.ScalarFunction;
 
 public class DolToEur extends ScalarFunction {
-    public double eval(double price) {
-        return price * 0.89;
+    public long eval(long price) {
+        return (long) (price * 0.89);
     }
 }
